first pass of refactor
