+++ lib/cluster/storage/analytics.js
@@ -76,7 +76,7 @@ module.exports = function(context) {
         shutdown: shutdown
     };
 
-    return require('./backends/elasticsearch')(context, index_name, 'analytics', '_id')
+    return require('./backends/elasticsearch').createStore(context, index_name, 'analytics', '_id')
         .then(function(elasticsearch) {
             logger.info("AnalyticsStorage: initializing");
             backend = elasticsearch;
