+++ lib/cluster/storage/backends/elasticsearch.js
@@ -4,14 +4,229 @@ var uuid = require('uuid');
 var Promise = require('bluebird');
 var _ = require('lodash');
 var fs = require('fs');
+var moment = require('moment');
 var filterResponse = require('../../../utils/elastic_utils').filterResponse;
 var warn = require('../../../utils/elastic_utils').warn;
 var elasticError = require('../../../utils/error_utils').elasticError;
-
+var buildQuery = require('../../../utils/elastic_utils').buildQuery;
 
 // Module to manage persistence in Elasticsearch.
 // All functions in this module return promises that must be resolved to get the final result.
-module.exports = function(context, index_name, record_type, id_field, bulk_size) {
+
+//TODO ID reader
+function oreader(context, client, opConfig, msg, logger) {
+    var query = {
+        index: opConfig.index,
+        size: msg.count ? msg.count : 200000,
+        body: {
+            query: {
+                wildcard: {
+                    _uid: msg.key
+                }
+            }
+        }
+    };
+    //Duplicate
+    return new Promise(function(resolve, reject) {
+        client.search(query)
+            .then(function(data) {
+                if (data._shards.failed > 0) {
+                    var reasons = _.uniq(_.flatMap(data._shards.failures, function(shard) {
+                        return shard.reason.type
+                    }));
+
+                    if (reasons.length > 1 || reasons[0] !== 'es_rejected_execution_exception') {
+                        var errorReason = reasons.join(' | ');
+                        logger.error('Not all shards returned successful, shard errors: ', errorReason);
+                        reject(errorReason)
+                    }
+                    else {
+                        // Spot to recurse in the future, will reject for now
+                        var errorReason = reasons.join(' | ');
+                        logger.error('Not all shards returned successful, shard errors: ', errorReason);
+                        reject(errorReason)
+                    }
+                }
+                else {
+                    if (opConfig.full_response) {
+                        resolve(data)
+                    }
+                    else {
+                        resolve(_.map(data.hits.hits, function(hit) {
+                            return hit._source
+                        }));
+                    }
+                }
+            })
+            .catch(function(err) {
+                var errMsg = elasticError(err);
+                logger.error(errMsg);
+                reject(errMsg)
+            });
+    })
+}
+
+
+//TODO date reader
+function reader(context, client, opConfig, msg, logger) {
+    return new Promise(function(resolve, reject) {
+        var query = buildQuery(opConfig, msg);
+
+        client.search(query).then(function(data) {
+            if (data._shards.failed > 0) {
+                var reasons = _.uniq(_.flatMap(data._shards.failures, function(shard) {
+                    return shard.reason.type
+                }));
+
+                if (reasons.length > 1 || reasons[0] !== 'es_rejected_execution_exception') {
+                    var errorReason = reasons.join(' | ');
+                    logger.error('Not all shards returned successful, shard errors: ', errorReason);
+                    reject(errorReason)
+                }
+                else {
+                    // Spot to recurse in the future, will reject for now
+                    var errorReason = reasons.join(' | ');
+                    logger.error('Not all shards returned successful, shard errors: ', errorReason);
+                    reject(errorReason)
+                }
+            }
+            else {
+                if (opConfig.full_response) {
+                    resolve(data)
+                }
+                else {
+                    resolve(_.map(data.hits.hits, function(hit) {
+                        return hit._source
+                    }));
+                }
+            }
+        }).catch(function(err) {
+            var errMsg = elasticError(err);
+            logger.error(errMsg);
+            reject(errMsg)
+        });
+    })
+}
+
+//TODO slicer getCount dateRange
+
+function queryCount(client, query, logger) {
+    return new Promise(function(resolve, reject) {
+        client.search(query)
+            .then(function(results) {
+                if (results._shards.failed > 0) {
+                    var reasons = _.uniq(_.flatMap(results._shards.failures, function(shard) {
+                        return shard.reason.type
+                    }));
+
+                    if (reasons.length > 1 || reasons[0] !== 'es_rejected_execution_exception') {
+                        var errorReason = reasons.join(' | ');
+                        logger.error(`Multiple elasticsearch errors occurred while determining count of slice: ${errorReason}`);
+                        reject(errorReason)
+                    }
+                    else {
+                        // Spot to recurse in the future, will reject for now
+                        var errorReason = reasons.join(' | ');
+                        logger.error(`elasticsearch error occurred while determining count of slice: ${errorReason}`);
+                        reject(errorReason)
+                    }
+                }
+                else {
+                    resolve(results.hits.total)
+                }
+            })
+            .catch(function(err) {
+                var errMsg = elasticError(err);
+                logger.error(`error while processing count of slice`, errMsg);
+                reject(errMsg)
+            });
+    })
+}
+
+
+//utility slicer
+
+function parseDate(date) {
+    var result;
+
+    if (moment(new Date(date)).isValid()) {
+        result = moment(new Date(date));
+    }
+    else {
+        var ms = dateMath.parse(date);
+        result = moment(ms)
+    }
+
+    return result;
+}
+
+// slicer get dates
+
+function autoDates(client, opConfig, query, givenDate, order, dateFormat, logger) {
+    return client.search(query)
+        .then(function(data) {
+            if (data.hits.hits.length === 0) {
+                return null
+            }
+
+            if (data.hits.hits[0] && data.hits.hits[0]._source[opConfig.date_field_name] === undefined) {
+                return Promise.reject(`date_field_name: "${opConfig.date_field_name}" for index: ${opConfig.index} does not exist`);
+            }
+
+            if (givenDate) {
+                return givenDate;
+            }
+
+            if (order === 'start') {
+                return parseDate(data.hits.hits[0]._source[opConfig.date_field_name])
+            }
+            else {
+                //end date is non-inclusive, adding 1s so range will cover it
+                var date = data.hits.hits[0]._source[opConfig.date_field_name];
+                var time = moment(date).add(1, 's');
+                return parseDate(time.format(dateFormat));
+            }
+        })
+        .catch(function(err) {
+            var errMsg = elasticError(err);
+            logger.error(errMsg);
+            return Promise.reject(errMsg);
+        });
+}
+
+function checkVersion(str) {
+    var num = Number(str.replace(/\./g, ''));
+    return num >= 210;
+}
+
+// date slicer
+function elasticsearchVersion(client, opConfig, logger) {
+    return client.cluster.stats({})
+        .then(function(data) {
+            var version = data.nodes.versions[0];
+
+            if (checkVersion(version)) {
+                return client.indices.getSettings({})
+                    .then(function(results) {
+                        if (results[opConfig.index]) {
+                            var data = results[opConfig.index].settings.index.max_result_window;
+                            var window = data ? data : 10000;
+
+                            logger.warn(`max_result_window for index: ${opConfig.index} is set at ${window} . On very large indices it is possible that a slice can not be divided to stay below this limit. If that occurs an error will be thrown by Elasticsearch and the slice can not be processed. Increasing max_result_window in the Elasticsearch index settings will resolve the problem.`);
+                        }
+                        else {
+                            return Promise.reject('index specified in reader does not exist')
+                        }
+                    }).catch(function(err) {
+                        var errMsg = elasticError(err);
+                        logger.error(errMsg);
+                        return Promise.reject(errMsg)
+                    })
+            }
+        });
+}
+
+function createStore(context, index_name, record_type, id_field, bulk_size) {
     var logger = context.foundation.makeLogger('elasticsearch_backend', 'elasticsearch_backend', {module: 'elasticsearch_backend'});
     var config = context.sysconfig.teraslice;
 
@@ -670,4 +885,65 @@ module.exports = function(context, index_name, record_type, id_field, bulk_size)
 
             });
     });
+}
+
+function putTemplate(client, template, name){
+    return client.indices.putTemplate({body: template, name: name})
+        .then(function(results) {
+            return results
+        })
+        .catch(function(err) {
+            var errMsg = elasticError(err);
+            return Promise.reject(errMsg)
+        })
+}
+
+function bulkSend(client, data, warning, logger){
+    return new Promise(function(resolve, reject) {
+        function sendData(client, data) {
+            client.bulk({body: data})
+                .then(function(results) {
+                    if (results.errors) {
+                        var response = filterResponse(logger, data, results);
+
+                        if (response.error) {
+                            reject(response.reason)
+                        }
+                        else {
+                            //may get doc already created error, if so just return
+                            if (response.data.length === 0) {
+                                resolve(results)
+                            }
+                            else {
+                                warning();
+                                setTimeout(function() {
+                                    sendData(client, response.data)
+                                }, 3000)
+                            }
+                        }
+                    }
+                    else {
+                        resolve(results)
+                    }
+                })
+                .catch(function(err) {
+                    var errMsg = elasticError(err);
+                    logger.error(`bulk sender error: ${errMsg}`);
+                    reject(`bulk sender error: ${errMsg}`);
+                })
+        }
+
+        sendData(client, data);
+    });
+}
+
+module.exports = {
+    createStore: createStore,
+    reader: reader,
+    oreader: oreader,
+    elasticsearchVersion: elasticsearchVersion,
+    autoDates: autoDates,
+    queryCount: queryCount,
+    putTemplate: putTemplate,
+    bulkSend: bulkSend
 };
