+++ lib/cluster/storage/jobs.js
@@ -92,7 +92,7 @@ module.exports = function(context, jobType) {
         jobArgs.push('job', 'ex_id')
     }
 
-    return require('./backends/elasticsearch').apply(null, jobArgs)
+    return require('./backends/elasticsearch').createStore.apply(null, jobArgs)
         .then(function(elasticsearch) {
             logger.info(`Initializing`);
             backend = elasticsearch;
