+++ lib/cluster/storage/state.js
@@ -102,7 +102,7 @@ module.exports = function(context) {
         shutdown: shutdown
     };
 
-    return require('./backends/elasticsearch')(context, index_name, 'state', '_id')
+    return require('./backends/elasticsearch').createStore(context, index_name, 'state', '_id')
         .then(function(elasticsearch) {
             logger.info("initializing");
             backend = elasticsearch;
