+++ lib/processors/elasticsearch_bulk.js
@@ -7,6 +7,7 @@ var getOpConfig = require('../utils/config').getOpConfig;
 var filterResponse = require('../utils/elastic_utils').filterResponse;
 var warn = require('../utils/elastic_utils').warn;
 var elasticError = require('../utils/error_utils').elasticError;
+var bulkSend = require('../cluster/storage/backends/elasticsearch').bulkSend;
 
 function newProcessor(context, opConfig, jobConfig) {
     var context = context;
@@ -83,52 +84,7 @@ function newProcessor(context, opConfig, jobConfig) {
 
     function send(client, data) {
         var warning = warn(logger, 'The elasticsearch cluster queues are overloaded, resubmitting failed queries from bulk');
-        let retryTimer = {start: 5000, limit: 10000};
-
-        return new Promise(function(resolve, reject) {
-            function sendData(client, data) {
-                client.bulk({body: data})
-                    .then(function(results) {
-                        if (results.errors) {
-                            var response = filterResponse(logger, data, results);
-
-                            if (response.error) {
-                                reject(response.reason)
-                            }
-                            else {
-                                //may get doc already created error, if so just return
-                                if (response.data.length === 0) {
-                                    resolve(results)
-                                }
-                                else {
-                                    let timer = Math.floor(Math.random() * (retryTimer.limit - retryTimer.start) + retryTimer.start);
-
-                                    if (retryTimer.limit < 60000) {
-                                        retryTimer.limit += 10000
-                                    }
-                                    if (retryTimer.start < 30000) {
-                                        retryTimer.start += 5000
-                                    }
-                                    warning();
-                                    setTimeout(function() {
-                                        sendData(client, response.data)
-                                    }, timer)
-                                }
-                            }
-                        }
-                        else {
-                            resolve(results)
-                        }
-                    })
-                    .catch(function(err) {
-                        var errMsg = elasticError(err);
-                        logger.error(`bulk sender error: ${errMsg}`);
-                        reject(`bulk sender error: ${errMsg}`);
-                    })
-            }
-
-            sendData(client, data);
-        });
+        return bulkSend(client, data, warning, logger);
     }
 
     function recursiveSend(client, dataArray) {
