+++ lib/readers/elasticsearch_date_range/reader.js
@@ -1,49 +1,8 @@
-var Promise = require('bluebird');
-var buildQuery = require('./../../utils/elastic_utils').buildQuery;
-var _ = require('lodash');
-var elasticError = require('../../utils/error_utils').elasticError;
-
+var elasticsearchReader = require('../../cluster/storage/backends/elasticsearch').reader;
 
 function newReader(context, opConfig, jobConfig, client) {
-
     return function(msg, logger) {
-        return new Promise(function(resolve, reject) {
-            var query = buildQuery(opConfig, msg);
-
-            client.search(query).then(function(data) {
-                if (data._shards.failed > 0) {
-                    var reasons = _.uniq(_.flatMap(data._shards.failures, function(shard) {
-                        return shard.reason.type
-                    }));
-
-                    if (reasons.length > 1 || reasons[0] !== 'es_rejected_execution_exception') {
-                        var errorReason = reasons.join(' | ');
-                        logger.error('Not all shards returned successful, shard errors: ', errorReason);
-                        reject(errorReason)
-                    }
-                    else {
-                        // Spot to recurse in the future, will reject for now
-                        var errorReason = reasons.join(' | ');
-                        logger.error('Not all shards returned successful, shard errors: ', errorReason);
-                        reject(errorReason)
-                    }
-                }
-                else {
-                    if (opConfig.full_response) {
-                        resolve(data)
-                    }
-                    else {
-                        resolve(_.map(data.hits.hits, function(hit) {
-                            return hit._source
-                        }));
-                    }
-                }
-            }).catch(function(err) {
-                var errMsg = elasticError(err);
-                logger.error(errMsg);
-                reject(errMsg)
-            });
-        })
+        return elasticsearchReader(context, client, opConfig, msg, logger)
     }
 }
