+++ lib/readers/elasticsearch_date_range/slicer.js
@@ -11,6 +11,9 @@ var dateFormatMS = require('./../../utils/elastic_utils').dateFormat;
 var dateFormatS = require('./../../utils/elastic_utils').dateFormatSeconds;
 var event = require('../../utils/events');
 var elasticError = require('../../utils/error_utils').elasticError;
+var elasticsearchVersion = require('../../cluster/storage/backends/elasticsearch').elasticsearchVersion;
+var autoDates = ('../../cluster/storage/backends/elasticsearch').autoDates;
+var queryCount = ('../../cluster/storage/backends/elasticsearch').queryCount;
 
 
 function newSlicer(context, opConfig, job, retryData, logger, client) {
@@ -24,39 +27,13 @@ function newSlicer(context, opConfig, job, retryData, logger, client) {
 
     function checkElasticsearch(client, opConfig, logger) {
         if (client) {
-            return client.cluster.stats({})
-                .then(function(data) {
-                    var version = data.nodes.versions[0];
-
-                    if (checkVersion(version)) {
-                        return client.indices.getSettings({})
-                            .then(function(results) {
-                                if (results[opConfig.index]) {
-                                    var data = results[opConfig.index].settings.index.max_result_window;
-                                    var window = data ? data : 10000;
-
-                                    logger.warn(`max_result_window for index: ${opConfig.index} is set at ${window} . On very large indices it is possible that a slice can not be divided to stay below this limit. If that occurs an error will be thrown by Elasticsearch and the slice can not be processed. Increasing max_result_window in the Elasticsearch index settings will resolve the problem.`);
-                                }
-                                else {
-                                    return Promise.reject('index specified in reader does not exist')
-                                }
-                            }).catch(function(err) {
-                                var errMsg = elasticError(err);
-                                logger.error(errMsg);
-                                return Promise.reject(errMsg)
-                            })
-                    }
-                });
+            return elasticsearchVersion(client, opConfig, logger)
         }
         else {
             return Promise.reject('no client is available for slicer')
         }
     }
 
-    function checkVersion(str) {
-        var num = Number(str.replace(/\./g, ''));
-        return num >= 210;
-    }
 
     function parseDate(date) {
         var result;
@@ -103,35 +80,7 @@ function newSlicer(context, opConfig, job, retryData, logger, client) {
         }
 
         //using this query to catch potential errors even if a date is given already
-        return client.search(query)
-            .then(function(data) {
-                if (data.hits.hits.length === 0) {
-                    return null
-                }
-
-                if (data.hits.hits[0] && data.hits.hits[0]._source[opConfig.date_field_name] === undefined) {
-                    return Promise.reject(`date_field_name: "${opConfig.date_field_name}" for index: ${opConfig.index} does not exist`);
-                }
-
-                if (givenDate) {
-                    return givenDate;
-                }
-
-                if (order === 'start') {
-                    return parseDate(data.hits.hits[0]._source[opConfig.date_field_name])
-                }
-                else {
-                    //end date is non-inclusive, adding 1s so range will cover it
-                    var date = data.hits.hits[0]._source[opConfig.date_field_name];
-                    var time = moment(date).add(1, 's');
-                    return parseDate(time.format(dateFormat));
-                }
-            }).catch(function(err) {
-                var errMsg = elasticError(err);
-                logger.error(errMsg);
-                return Promise.reject(errMsg);
-            });
-
+        return autoDates(client, opConfig, query, givenDate, order, dateFormat, logger);
     }
 
     function updateJob(opConfig, dates) {
@@ -181,7 +130,7 @@ function newSlicer(context, opConfig, job, retryData, logger, client) {
         }
     }
 
-    function determineSlice(client, config, dateParams, slicer_id, isExpandedSlice) {
+    function determineSlice(config, dateParams, slicer_id, isExpandedSlice) {
         return getCount(config, dateParams).then(function(count) {
             if (count > dateParams.size) {
                 //if size is to big after increasing slice, use alternative division behavior
@@ -195,7 +144,7 @@ function newSlicer(context, opConfig, job, retryData, logger, client) {
                     clonedParams.end = moment(clonedParams.start).add(diff, time_resolution);
 
                     //return the zero range start with the correct end
-                    return Promise.resolve(determineSlice(client, config, clonedParams, slicer_id, false))
+                    return Promise.resolve(determineSlice(config, clonedParams, slicer_id, false))
                         .then(function(recursedData) {
                             return {start: dateParams.start, end: recursedData.end, count: recursedData.count};
                         });
@@ -214,7 +163,7 @@ function newSlicer(context, opConfig, job, retryData, logger, client) {
                         dateParams.end = newEnd;
                         event.emit('slicer:slice:recursion');
                         logger.trace(`slicer: ${slicer_id} is recursing ${JSON.stringify(dateParams)}`);
-                        return determineSlice(client, config, dateParams, slicer_id, isExpandedSlice)
+                        return determineSlice(config, dateParams, slicer_id, isExpandedSlice)
                     }
                 }
             }
@@ -231,7 +180,7 @@ function newSlicer(context, opConfig, job, retryData, logger, client) {
                         var newEnd = moment(dateParams.end).add(dateParams.interval[0], dateParams.interval[1]);
                         dateParams.end = newEnd;
                         event.emit('slicer:slice:range_expansion');
-                        return determineSlice(client, config, dateParams, slicer_id, true)
+                        return determineSlice(config, dateParams, slicer_id, true)
                     }
                 }
                 else {
@@ -320,7 +269,7 @@ function newSlicer(context, opConfig, job, retryData, logger, client) {
                 return null;
             }
             else {
-                return determineSlice(client, opConfig, dateParams, slicer_id, false)
+                return determineSlice(opConfig, dateParams, slicer_id, false)
                     .then(function(data) {
                         dateParams.start = data.end;
 
@@ -441,7 +390,7 @@ function newSlicer(context, opConfig, job, retryData, logger, client) {
                 return null;
             }
             else {
-                return determineSlice(client, opConfig, dateParams, slicer_id, false)
+                return determineSlice(opConfig, dateParams, slicer_id, false)
                     .then(function(data) {
                             dateParams.start = data.end;
 
@@ -491,36 +440,7 @@ function newSlicer(context, opConfig, job, retryData, logger, client) {
         var query = buildQuery(opConfig, range);
 
         //duplication
-        return new Promise(function(resolve, reject) {
-            client.search(query)
-                .then(function(results) {
-                    if (results._shards.failed > 0) {
-                        var reasons = _.uniq(_.flatMap(results._shards.failures, function(shard) {
-                            return shard.reason.type
-                        }));
-
-                        if (reasons.length > 1 || reasons[0] !== 'es_rejected_execution_exception') {
-                            var errorReason = reasons.join(' | ');
-                            logger.error(`Multiple elasticsearch errors occurred while determining count of slice: ${errorReason}`);
-                            reject(errorReason)
-                        }
-                        else {
-                            // Spot to recurse in the future, will reject for now
-                            var errorReason = reasons.join(' | ');
-                            logger.error(`elasticsearch error occurred while determining count of slice: ${errorReason}`);
-                            reject(errorReason)
-                        }
-                    }
-                    else {
-                        resolve(results.hits.total)
-                    }
-                })
-                .catch(function(err) {
-                    var errMsg = elasticError(err);
-                    logger.error(`error while processing count of slice`, errMsg);
-                    reject(errMsg)
-                });
-        })
+        return queryCount(client, query, logger);
     }
 
     function getInterval(opConfig, esDates) {
