+++ lib/readers/id_reader.js
@@ -1,10 +1,9 @@
 'use strict';
 
-var Promise = require('bluebird');
 var _ = require('lodash');
 var getClient = require('../utils/config').getClient;
 var getOpConfig = require('../utils/config').getOpConfig;
-var elasticError = require('../utils/error_utils').elasticError;
+var elasticsearchReader = require('../cluster/storage/backends/elasticsearch').oreader;
 
 
 var parallelSlicers = true;
@@ -19,59 +18,7 @@ function newSlicer(context, job, retryData, slicerAnalytics, logger) {
 
 function newReader(context, opConfig, jobConfig) {
     var client = getClient(context, opConfig, 'elasticsearch');
-
-    return function(msg, logger) {
-        var query = {
-            index: opConfig.index,
-            size: msg.count ? msg.count : 200000,
-            body: {
-                query: {
-                    wildcard: {
-                        _uid: msg.key
-                    }
-                }
-            }
-        };
-        //Duplicate
-        return new Promise(function(resolve, reject) {
-            client.search(query)
-                .then(function(data) {
-                    if (data._shards.failed > 0) {
-                        var reasons = _.uniq(_.flatMap(data._shards.failures, function(shard) {
-                            return shard.reason.type
-                        }));
-
-                        if (reasons.length > 1 || reasons[0] !== 'es_rejected_execution_exception') {
-                            var errorReason = reasons.join(' | ');
-                            logger.error('Not all shards returned successful, shard errors: ', errorReason);
-                            reject(errorReason)
-                        }
-                        else {
-                            // Spot to recurse in the future, will reject for now
-                            var errorReason = reasons.join(' | ');
-                            logger.error('Not all shards returned successful, shard errors: ', errorReason);
-                            reject(errorReason)
-                        }
-                    }
-                    else {
-                        if (opConfig.full_response) {
-                            resolve(data)
-                        }
-                        else {
-                            resolve(_.map(data.hits.hits, function(hit) {
-                                return hit._source
-                            }));
-                        }
-                    }
-                })
-                .catch(function(err) {
-                    var errMsg = elasticError(err);
-                    logger.error(errMsg);
-                    reject(errMsg)
-                });
-        })
-    }
-
+    return elasticsearchReader(context, client, opConfig, logger);
 }
 
 function schema() {
