+++ lib/readers/id_slicer.js
@@ -3,6 +3,7 @@
 var elasticError = require('../utils/error_utils').elasticError;
 var getKeyArray = require('../utils/id_utils').getKeyArray;
 var buildQuery = require('../utils/elastic_utils').buildQuery;
+var queryCount = ('../../cluster/storage/backends/elasticsearch').queryCount;
 
 var _ = require('lodash');
 
@@ -14,36 +15,7 @@ module.exports = function(client, job, opConfig, logger, retryData, range) {
     var keyArray = opConfig.key_range ? opConfig.key_range : baseKeyArray.slice();
 
     function getCountForKey(query) {
-        return new Promise(function(resolve, reject) {
-            client.search(query)
-                .then(function(results) {
-                    if (results._shards.failed > 0) {
-                        var reasons = _.uniq(_.flatMap(results._shards.failures, function(shard) {
-                            return shard.reason.type
-                        }));
-
-                        if (reasons.length > 1 || reasons[0] !== 'es_rejected_execution_exception') {
-                            var errorReason = reasons.join(' | ');
-                            logger.error(errorReason);
-                            reject(errorReason)
-                        }
-                        else {
-                            // Spot to recurse in the future, will reject for now
-                            var errorReason = reasons.join(' | ');
-                            logger.error(errorReason);
-                            reject(errorReason)
-                        }
-                    }
-                    else {
-                        resolve(results.hits.total)
-                    }
-                })
-                .catch(function(err) {
-                    var errMsg = elasticError(err);
-                    logger.error(errMsg);
-                    reject(errMsg)
-                });
-        })
+        return queryCount(client, query, logger);
     }
 
     function determineKeySlice(generator, closePath, rangeObj) {
